fix: 解决单位转换导致的内存限制问题 (#4055)
Refs #4040

+++ backend/app/dto/container.go
@@ -38,8 +38,8 @@ type ContainerInfo struct {
 }
 
 type ResourceLimit struct {
-	CPU    int `json:"cpu"`
-	Memory int `json:"memory"`
+	CPU    int    `json:"cpu"`
+	Memory uint64 `json:"memory"`
 }
 
 type ContainerOperate struct {
